+++ src/ui/events.ts
@@ -3,14 +3,15 @@ import {Event} from '../util/evented';
 import DOM from '../util/dom';
 import Point from '@mapbox/point-geometry';
 import {extend} from '../util/util';
+import type {MapGeoJSONFeature} from '../util/vectortile_to_geojson';
 
 import type Map from './map';
 import type LngLat from '../geo/lng_lat';
 import {SourceSpecification} from '@maplibre/maplibre-gl-style-spec';
 
-export type MapLayerMouseEvent = MapMouseEvent & { features?: GeoJSON.Feature[] };
+export type MapLayerMouseEvent = MapMouseEvent & { features?: MapGeoJSONFeature[] };
 
-export type MapLayerTouchEvent = MapTouchEvent & { features?: GeoJSON.Feature[] };
+export type MapLayerTouchEvent = MapTouchEvent & { features?: MapGeoJSONFeature[] };
 
 export type MapSourceDataType = 'content' | 'metadata';
 
@@ -31,7 +32,7 @@ export type MapLayerEventType = {
     touchcancel: MapLayerTouchEvent;
 };
 
-export interface MapLibreEvent<TOrig = undefined> {
+export interface MapLibreEvent<TOrig = unknown> {
     type: string;
     target: Map;
     originalEvent: TOrig;